requirements.txt
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==41.0.4
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.24.4
packaging==25.0
pandas==2.1.4
platformdirs==3.11.0
pluggy==1.6.0
psycopg2-binary==2.9.7
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==23.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.4
six==1.17.0
snowflake-connector-python==3.4.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
tomlkit==0.13.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
Werkzeug==3.1.3
